@@ -4,23 +4,27 @@
   "private": true,
   "homepage": "/resume",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.4.3",
-    "bootstrap": "^5.2.3",
-    "bootstrap-icons": "^1.10.3",
-    "framer-motion": "^9.0.4",
+    "@testing-library/jest-dom": "^6.6.0",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/user-event": "^14.5.2",
+    "bootstrap": "^5.3.3",
+    "bootstrap-icons": "^1.11.3",
+    "framer-motion": "^11.15.0",
     "lodash": "^4.17.21",
-    "node-sass": "^8.0.0",
-    "react": "^18.2.0",
-    "react-bootstrap": "^2.7.2",
-    "react-bootstrap-icons": "^1.10.2",
-    "react-dom": "^18.2.0",
-    "react-icons": "^4.7.1",
-    "react-router-dom": "^6.8.1",
+    "react": "^19.1.0",
+    "react-bootstrap": "^2.10.6",
+    "react-bootstrap-icons": "^1.11.4",
+    "react-dom": "^19.1.0",
+    "react-icons": "^5.4.0",
+    "react-intersection-observer": "^9.16.0",
+    "react-router-dom": "^7.6.3",
     "react-scripts": "5.0.1",
-    "react-vertical-timeline-component": "^3.6.0",
-    "web-vitals": "^3.1.1"
+    "react-vertical-timeline-component": "^3.5.3",
+    "sass": "^1.89.2",
+    "web-vitals": "^4.2.4"
+  },
+  "overrides": {
+    "react-intersection-observer": "^9.16.0"
   },
   "scripts": {
     "start": "react-scripts start",
